PHASE1.md and PHASE2.md; freeze experiments.md; theory revision 2
DELIVERABLE 1 -- two self-contained source-of-truth documents at the repo root.
Every number carries its producing script and output file inline, beside the
number, plus spread, seed/split count, the control it is measured against, and
the threshold it was judged by. No model number appears without a control.

  PHASE1.md   the audit. Instrument validity, non-degeneracy, A1, A2, A3,
              mechanism profiling, E5.1 and E4.1, then a separated section 9 for
              what is not established.
  PHASE2.md   downstream. Experiments 2.1-2.7, then section 5 for scope limits,
              section 6 for discrepancies found, section 7 for what supersedes
              what.

Reconciled, not concatenated. Five discrepancies found and resolved, listed in
PHASE2.md section 6. The load-bearing one: four phase_2 documents state "the
tier-0 file still carries the bug" about TargetedImitator.inner_jacobian, which
became false when it was fixed at source. Numbers are unaffected (bit-identical,
verified), so the record documents are left as history and a correction banner
was added to phase_2/README.md, which is navigation rather than record.

Also recorded: no file named PHASE2_RESULTS.md has ever existed, in this repo or
in any brief that cited it. LOCALISATION.md section 0 records the same confusion
arising once before.

DELIVERABLE 2

experiments.md FROZEN with a header stating it is a dated pre-registration and
is never edited, including its stale inline Status lines.

experiments/DEVIATIONS.md is new: one row per experiment against four outcomes
-- RUN, RUN-D, NOT RUN, FIRED. Tally: 5 run as specified, 10 run with a stated
deviation, 12 not run, 3 falsifiers fired.

E5.2 is written the way it was asked to be written. The prediction was stated in
advance and every axis it names runs the other way -- TabPFN is worse than
TabICL on A1 (1.312 vs 0.580) and A2 (0.404 vs 0.177), and row scaling was not
recovered (residual 0.616 against 1.5e-15 on the NW control). But no registered
run exists; the numbers came from exp3_reprobe.py and exp2_ambient_profiling.py.
The log states explicitly what the paper may claim (the audit's measurements are
inconsistent with the Bilos prediction, stated in advance) and what it may not
(a pre-registered test falsified it).

theory_and_claims.md bumped to revision 2 with an amendment log at section 5,
amendments recorded rather than applied silently:
  A1  N4 proven and without instrument power against the failure that occurred;
      cv_s is permutation-invariant, so it holds for both models while R^2 is
      0.0199 and 0.1065. Both facts stand.
  A2  N5's exclusion inference softened to what section 7 measures -- neither
      row nor column scaling fits, residuals 0.476-0.616 against the NW control
      at 1.5e-15. E2.6 was not run, so no calibrated tolerance exists and the
      headline scatter cannot be drawn.
  A3  N7 reserved, DERIVATION PENDING, with the empirical fact it must explain
      and the three things already known about it. Not invented.
  A4  C-X2 promoted from claim to result, instantiated by E4.1.
Amended statements of C-N4, C-N5, C-E6, C-X2 and a new pending C-N7 follow the log.

experiments/README.md now points at the root documents.

+++ models/interp/adapters/tabicl_v2.py
@@ -81,7 +81,7 @@ class TabICLv2Interp(IclStackInterp):
             levels if levels is not None else np.linspace(0.005, 0.995, 199), dtype=float
         )
         q = self.estimator.predict(
-            np.asarray(X_test, dtype=np.float64),
+            np.asarray(X_test),
             output_type="quantiles",
             alphas=[float(a) for a in levels],
         )
